Minor refactor and cleanup to improve testability
